Refactored project. Added new regex in phone validator.
