+++ src/main/java/ru/javaluxurywatches/controller/user/UserController.java
@@ -10,9 +10,9 @@ import ru.javaluxurywatches.service.user.UserService;
 @RequestMapping("/users")
 public class UserController {
 
-    @Autowired
     private final UserService userService;
 
+    @Autowired
     public UserController(UserService userService) {
         this.userService = userService;
     }
