+++ src/main/java/ru/javaluxurywatches/interceptor/ProfileInterceptor.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Component;
 import org.springframework.web.servlet.ModelAndView;
 import org.springframework.web.servlet.handler.HandlerInterceptorAdapter;
 import ru.javaluxurywatches.model.user.User;
-import ru.javaluxurywatches.service.user.UserDetailService;
+import ru.javaluxurywatches.service.user.UserEntityManager;
 import ru.javaluxurywatches.service.user.UserService;
 
 import javax.servlet.http.HttpServletRequest;
@@ -16,11 +16,11 @@ import javax.servlet.http.HttpServletResponse;
 @Component
 public class ProfileInterceptor extends HandlerInterceptorAdapter {
 
-    private final UserDetailService userDetailService;
+    private final UserEntityManager userDetailService;
     private final UserService userService;
 
     @Autowired
-    public ProfileInterceptor(UserDetailService userDetailService,
+    public ProfileInterceptor(UserEntityManager userDetailService,
                               UserService userService) {
         this.userDetailService = userDetailService;
         this.userService = userService;
@@ -31,9 +31,11 @@ public class ProfileInterceptor extends HandlerInterceptorAdapter {
                            HttpServletResponse response,
                            Object handler,
                            ModelAndView modelAndView) throws Exception {
-        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
-        User user = userService.findByLogin(auth.getName());
-        modelAndView.addObject("userDetail", user.getUserDetail());
-        super.postHandle(request, response, handler, modelAndView);
+        if (modelAndView != null) {
+            Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+            User user = userService.findByLogin(auth.getName());
+            modelAndView.addObject("userDetail", user.getUserDetail());
+            super.postHandle(request, response, handler, modelAndView);
+        }
     }
 }
