+++ src/main/resources/import.sql
@@ -1,8 +1,8 @@
 INSERT INTO "user_detail" (gender,address,city,country,postcode,day_of_birth) VALUES ('Male', 'Sumska 24', 'Kharkiv', 'Ukrain', 600100, now());
 INSERT INTO "user_detail" (gender,address,city,country,postcode,day_of_birth) VALUES ('Female', 'Arbat 25', 'Moscow', 'Russia', 601044, now());
 
-INSERT INTO "user" (login, password, first_name, last_name, email, phone, is_active, user_detail_id) VALUES ('admin', '123456', 'walter', 'white', 'admin@lux.com','3053852214', true, 1);
-INSERT INTO "user" (login, password, first_name, last_name, email, phone, is_active, user_detail_id) VALUES ('user', '123456', 'jesse', 'pinkman','user@lux.com','3043862218', true, 2);
+INSERT INTO "user" (login, password, first_name, last_name, email, phone, is_active, user_detail_id) VALUES ('admin', '123456', 'walter', 'white', 'admin@lux.com','+7-900-800-70-60', true, 1);
+INSERT INTO "user" (login, password, first_name, last_name, email, phone, is_active, user_detail_id) VALUES ('user', '123456', 'jesse', 'pinkman','user@lux.com','88006005530', true, 2);
 
 INSERT INTO "role" (name) VALUES ('ROLE_ADMIN');
 INSERT INTO "role" (name) VALUES ('ROLE_USER');
